feat: Extract `BlockTest` from `OtherTests` and write more block tests
